requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.2.10
async-timeout==4.0.2
attrs==22.2.0
black==23.1.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
Django==3.1
djangorestframework==3.14.0
dnspython==2.3.0
drf-pydantic==0.3.0
email-validator==1.3.1
et-xmlfile==1.1.0
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.24.2
openpyxl==3.1.2
packaging==23.0
pandas==1.5.3
pathspec==0.11.0
platformdirs==3.1.1
pydantic==1.10.6
python-dateutil==2.8.2
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.5.0
yarl==1.8.2
